+++ check_log.py
@@ -336,7 +336,8 @@ def fetch_log(args_array, **kwargs):
 
         # If file is closed, open up next one.
         if log_file.closed:
-            log_file = open(x, "r")
+            log_file = gen_libs.openfile(x, "r")
+            #log_file = open(x, "r")
 
         log_array.extend(gen_libs.get_data(log_file))
         log_file.close()