+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Changed
 - open_log:  Replace open call with gen_libs.openfile call.
+- find_marker:  Replace open calls with gen_libs.openfile calls.
 
 
 ## [2.2.0] - 2019-06-11