+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - find_marker:  Replace open calls with gen_libs.openfile calls.
 - get_ignore_msgs:  Replace open call with gen_libs.openfile call.
 - fetch_log:  Replace open call with gen_libs.openfile call.
+- get_filter_data:  Replace open call with gen_libs.openfile call.
 
 
 ## [2.2.0] - 2019-06-11